+++ ChangeLog.md
@@ -25,6 +25,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - (Import) Better error handling when textures are missing
 - (Export) Remember destination path when exporting individual GameObjects from menu
 - (Export) Vertical texture transform offset is correct now
+- Improved relative file path handling on platforms with non-forward slash directory separator (Windows)
 
 ## [4.6.0] - 2022-02-23
 ### Added